Adding active status to currently selected navigation item
Currently, when select a item in the top navigation bar, the item has only
text highlighted, with background color stay the same. When click on somewhere
else on the screen, the highlight disappeared. This is because we did't apply a
`active` class to the selected item.

This PR fixes the issue.