Allow Gulp Watch via docker compose
It'd be nice to have a docker-compose command which enables a
development environment to be running the way we want a development
environment to be running.  It turns out that the
docker-compose.override.yml file provides a set of "defaults".  Thus, if
an alternative file is specified, the override file will be ignored
entirely.