Convert to extends notation and use db.env
In looking up how folks are deploying things in production vs.
development, and discovered that the generally accepted practice is to
have a common.yml file and extend off of that.

I also abstracted out the database connection information into a db.env
file.  The env_file mechanism is designed for this exact kind of thing.