TEAMFOUR-154 Admin register clusters
When an admin logs in, if no clusters are currently registered, they
should be taken to an overlay where they can add clusters.  Once they
have added atleast one cluster they should see a list of the clusters
they've added.  This table should include a remove button for each
cluster so it can be removed from the list.