Adding jQuery for karma tests
Because jqLite is so lacking in selectors, writing unittests is much
harder than it needs to be without jQuery.  BUT... we don't want to add
jQuery to our production product.  So, we can just add it to the
devDependencies in bower, and then unittests will need to explicitly
wrap angular.element objects with $().