Ensure the dist folder is clear before building
We had an issue with the correct version of the dist folder not making its way into the dellivered image.

It's an artiufact of the build process and should be cleaned up before each build/publish.

It's useful to have it around AFTER the build so we know the state of it as a result of the build.