473 Removing the last HCF cluster demands another
Previously, if you remove the last HCF cluster from the services
registration page, the UI would be left in a weird state because you
don't have any clusters registered.

We don't allow that behavior upon login because it's ridiculous.

So, FOR NOW, let's just induce a refresh if the last cluster is removed.