create vmware docker-machine
Since many of us are using macs, we don't have docker running locally.
Now with the change to start using UCP (HCP?) (PCP?) we will need to
have VMWare Fusion running because apparently, ?CP won't run on vBox.
This means that it would make a LOT more sense to run your
docker-machine on VMWare instead of vBox.

The included script facilitates that.