T4-518 - remove deprecated elements of the console
This removes several, deprecated elements of the Stackato Console, as they are no longer necessary:

- nodejs server
- mock auth server
- mock api server
- mock cloud foundry service
- mysql db server

BREAKING CHANGE:
In the interests of getting a small footprint application stack working ASAP, I've temporarily broken the GitHub OAuth mechanism, as that currently resides inside the NodeJS API service. Once this PR merge, we only need to combine a couple of repos, and work can stanrt on the automated build via Concourse - we UNBLOCK that effort. OAuth willl now be done by the proxy, and a PR that implements that will be quickly forthcoming.