Change GitHub OAuth as a service
Currently GitHub authentication is an angular directive which put a button on screen to trigger GiHub OAuth workflow. This PR change it to a angular service so that we can combine GitHub OAuth workflow as part of the async operations after user click 'Next' button. If user denied the request of OAuth, it won't take user to next step in add-app workflow.