HSC-1136 Increate timeout value for postfllight job
I misunderstood the intent/meaning of the timeout attribute for a job component. I was under the impression that it was specific to a given attempted run of the job, and subject to 1-100 retries if necessary. Actually, it’s the about the total amount of time that might pass before the job is terminated and not allowed to run again.

This PR fixes this with a value of 30 minnutes vs 5 minutes.