Revert back to the use of ENTRYPOINT
As Neil corrrectly pointed out, the behavior of CMD vs ENTRYPOINT is different. Given I’m not seeing the logs get sent to FlightRecorder, yet I can do so manually, the likely culprit is that the default stolon ENTRYPOINT is boing run and my CMD is being run afterwards. Oops. This change (hopefully) fixes that.