HSC-1342: Notification taget list contains stale values after recreating pipeline (#876)
* HSC-1342: Notification taget list contains stale values after recreating pipeline
- Digest was crashing and therefor project was not cleared when deleted
- Also fixed position of content in notification wizard

* Additional fix - Ensure we clear stale data
Delivery pipeline project was never updated after project was deleted +
recreated. This was due to the application project being updated after
a delivery pipeline watch was triggered.
The fix ensures that any stale data is gone, which stops any update request
using stale data and then ensures the delivery pipeline watch is triggered
when a new pipeline is created.