Removed pattern at risk of optimisation (#1145)
Some services containing extension points are not referenced anywhere. In order
for it to register with an extension point this needs to happen. Previously
this was done in a run block with no usage of said service. This might be at
risk of optimisation so ensure we explicitly call something in the service instead.