marketplace: cf endpoint on card when multiple cf connected
When multiple CF were connected the user could end up with multiple services
with the same name without being able to identify the source of that.

Now when multiple CF endpoints are connected a CF endpoint is shown on
each service card so the user can identify where it comes from.

Signed-off-by: Vítor Avelino <vavelino@suse.com>