update kubernetes deploy doc to reference uaa.endpoint
noticed that in this commit which updates the stratos console uaa configuration to be simpler and leverage "uaa.endpoint" that the README was using an example referencing "uaa.url" (which I don't think is a param reference? could be wrong?) and still including uaa.port, uaa.host, etc., though it appears that was all replaced by `uaa.endpoint`

no rush; would appreciate if someone double-checked this, I didn't completely confirm or test this, to be honest. I'll look more into it this evening after work. apologies if I'm dead wrong.

thanks for the time